requirements.txt
feedparser==6.0.11
requests==2.31.0
python-dotenv==1.0.1
schedule==1.2.2
pandas==2.2.2
jinja2==3.1.4
anthropic==0.28.0
